services/resource-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic-settings>=2.0
asyncpg
fastcrud
httpx
pika